fix: proxy group ipv6 test failed
