+++ src/components/IPv6Support.tsx
@@ -1,8 +1,7 @@
 import { useProxies } from '~/signals'
-import { proxyIPv6SupportMap } from '~/signals/ipv6'
 
 export const IPv6Support = (props: { name?: string }) => {
-  const { getNowProxyNodeName } = useProxies()
+  const { getNowProxyNodeName, proxyIPv6SupportMap } = useProxies()
 
   const support = createMemo(() => {
     return proxyIPv6SupportMap()[getNowProxyNodeName(props.name || '')]
